stop distinguishing CPython & PyPy for language-level >=3.9
numpy has been passing the test suite on PyPy 3.8 & 3.9 since at least 1.19.
There are no wheels for PyPy 3.9 until numpy 1.25, but neither are there
official wheels for loongarch64 etc., so we should reflect that PyPy is
supported.

Future PyPy versions will only be more compatible with the respective CPython
version, so it's not a stretch to mirror the constraints we have for (say)
CPython 3.11 to those for a future PyPy 3.11.